refactor(controller): kill the bridge — compute owns the one-active-model surface
Stage A of the engine-registry plan. The ComputeBridge adapter is gone as a
separate AppContext member: the recipe->launch translation and the fixed "llm"
instance convention now live in compute/active-model.ts, created inside
makeCompute, and every former context.bridge consumer reads
context.compute.model. The four legacy lifecycle routes (/launch/:recipeId,
/launch/:recipeId/cancel, /evict, /wait-ready) moved from engines into
compute/routes.ts with identical wire shapes, and the crash-loop budget moved
with them; engines/lifecycle-routes.ts is deleted. Verified live: scratch boot,
404/evict/wait-ready shapes unchanged, three spawn failures then the 429
budget block, /status reporting launching + launch_failures.
