+++ controller/src/modules/engines/observed-process.ts
@@ -4,8 +4,8 @@ import { observeControllerFunction } from "../../core/function-observability";
 export const createGetObservedProcess =
   (
     context: AppContext,
-  ): ((label: string) => ReturnType<AppContext["bridge"]["findInferenceProcess"]>) =>
+  ): ((label: string) => ReturnType<AppContext["compute"]["model"]["findInferenceProcess"]>) =>
   (label: string) =>
     observeControllerFunction(context, `${label}.getCurrentProcess`, () =>
-      context.bridge.findInferenceProcess(),
+      context.compute.model.findInferenceProcess(),
     );
