+++ controller/src/modules/proxy/openai-routes.ts
@@ -157,7 +157,7 @@ export const registerOpenAIRoutes = defineRoutes((app, context) => {
     requestedModel: string | null,
     sourceHeader: string | null,
   ): Effect.Effect<ModelNotRunningError | null, unknown> =>
-    context.bridge.findInferenceProcess().pipe(
+    context.compute.model.findInferenceProcess().pipe(
       Effect.map((current) => {
         const matches =
           current && isRecipeRunning(matchedRecipe, current, { allowEitherPathContains: true });
